# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Star Wars API",
        "description": "API documentation for Star Wars app.",
        "contact": {
            "email": "tkisiel5w4@yahoo.com"
        },
        "license": {
            "name": "MIT"
        },
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://localhost/",
            "description": "Star Wars API Server"
        }
    ],
    "paths": {
...
        "/api/login": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "authenticate existing user",
                "operationId": "App\\Http\\Controllers\\Auth\\LoginController::__invoke",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Returns new token when valid credentials are provided",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SignInSuccessResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "There is required to provide correct email and password to get token",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SignInFailedResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "The given data was invalid.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnprocessableResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/user/profile": {
            "put": {
                "tags": [
                    "auth"
                ],
                "summary": "update user profile data",
                "operationId": "App\\Http\\Controllers\\Auth\\ProfileController::__invoke",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProfileUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Returns acknowledge when the user profile has been successfully updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProfileUpdateSuccessResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "There is required to provide correct api token to see this resources",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthenticatedResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "The given data was invalid.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnprocessableResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/register": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "create new user account",
                "operationId": "App\\Http\\Controllers\\Auth\\RegisterController::__invoke",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RegisterRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "New user was successfully created.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RegisterCreatedResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "The given data was invalid.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnprocessableResponse"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "User": {
                "title": "User model",
                "description": "User model",
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "User id",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "hero_id": {
                        "title": "hero_id",
                        "description": "Correlated star wars hero's id",
                        "type": "integer",
                        "format": "int64",
                        "example": 23
                    },
                    "email": {
                        "title": "email",
                        "description": "Email address of the user",
                        "type": "string",
                        "example": "john.doe@example.com"
                    },
                    "created_at": {
                        "title": "created_at",
                        "description": "User creation date",
                        "type": "string",
                        "format": "datetime",
                        "example": "2021-05-19 15:30:12"
                    },
                    "updated_at": {
                        "title": "updated_at",
                        "description": "The last user edition date",
                        "type": "string",
                        "format": "datetime",
                        "example": "2021-05-19 15:30:12"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "User"
                }
            },
            "LoginRequest": {
                "title": "Login request",
                "description": "Login request body data",
                "required": [
                    "email",
                    "password"
                ],
                "properties": {
                    "email": {
                        "title": "email",
                        "description": "Email address to the user's account.",
                        "type": "string",
                        "example": "john.doe@example.com"
                    },
                    "password": {
                        "title": "password",
                        "description": "Password to the user's account.",
                        "type": "string",
                        "example": "Vertex25!"
                    }
                },
                "type": "object"
            },
            "ProfileUpdateRequest": {
                "title": "Profile update request",
                "description": "Profile update request body data",
                "required": [
                    "email"
                ],
                "properties": {
                    "email": {
                        "title": "email",
                        "description": "New email address to the user's account.",
                        "type": "string",
                        "example": "jane.doe@example.com"
                    }
                },
                "type": "object"
            },
            "RegisterRequest": {
                "title": "Registration request",
                "description": "Registration request body data",
                "required": [
                    "email",
                    "password",
                    "password_confirmation"
                ],
                "properties": {
                    "email": {
                        "title": "email",
                        "description": "Email address for the user being created.",
                        "type": "string",
                        "example": "john.doe@example.com"
                    },
                    "password": {
                        "title": "password",
                        "description": "Password for the user being created.",
                        "type": "string",
                        "example": "Vertex25!"
                    },
                    "password_confirmation": {
                        "title": "password",
                        "description": "Password confirmation for the user being created.",
                        "type": "string",
                        "example": "Vertex25!"
                    }
                },
                "type": "object"
            },
...
            "ProfileUpdateSuccessResponse": {
                "title": "Sign in success response",
                "description": "When email and password are valid",
                "properties": {
                    "message": {
                        "title": "message",
                        "type": "string",
                        "format": "string",
                        "example": "Great! Your profile has been successfully updated."
                    }
                },
                "type": "object"
            },
            "RegisterCreatedResponse": {
                "title": "Registration created response",
                "description": "Registration response after user being created.",
                "properties": {
                    "message": {
                        "title": "message",
                        "type": "string",
                        "example": "Your account has been successfully created! Now You can sign in via login endpoint."
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    }
                },
                "type": "object"
            },
            "SignInFailedResponse": {
                "title": "Sign in failed response",
                "description": "When email and password are invalid or missing",
                "properties": {
                    "message": {
                        "title": "message",
                        "type": "string",
                        "example": "Unfortunately! The given credentials don\\'t match any existing user."
                    }
                },
                "type": "object"
            },
            "SignInSuccessResponse": {
                "title": "Sign in success response",
                "description": "When email and password are valid",
                "properties": {
                    "message": {
                        "title": "message",
                        "type": "string",
                        "format": "string",
                        "example": "You have been logged in successfully. You can now access your data via the bearer API token."
                    },
                    "token": {
                        "title": "token",
                        "description": "Bearer token that you can use to get to secure endpoints",
                        "type": "string",
                        "format": "string",
                        "example": "ogDCM1z1vrwHDhRwMEBby5AlmvY0WrmOGkfN7Rn0LyckOkbeEI4l9IanRwBrKdCl"
                    }
                },
                "type": "object"
            },
            "UnauthenticatedResponse": {
                "title": "Unauthenticated response",
                "description": "When token is invalid or missing",
                "properties": {
                    "message": {
                        "title": "message",
                        "type": "string",
                        "example": "You're unauthenticated because of wrong or not provided api token. Please make sure that your token is correct or sign in again to get new one."
                    }
                },
                "type": "object"
            },
            "UnprocessableResponse": {
                "title": "Unprocessable response",
                "description": "When given data was invalid.",
                "properties": {
                    "message": {
                        "title": "message",
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "title": "errors",
                        "description": "Description of invalid entites",
                        "properties": {
                            "email": {
                                "type": "object",
                                "example": [
                                    "The email field is required."
                                ]
                            },
                            "password": {
                                "type": "object",
                                "example": [
                                    "The password must be at least 8 characters.",
                                    "Password is too weak. At least one letter, number and special character are required."
                                ]
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        []
    ]
}